refactor: componentsフォルダーをembedにリネーム
