+++ packages/zenn-embed-components/src/embed/gist/EmbedGistLoading.tsx
