+++ packages/zenn-embed-components/src/embed/github/EmbedGitHubError.tsx
