+++ packages/zenn-embed-components/src/embed/link-card/EmbedLinkCardPlaceholder.tsx
