+++ packages/zenn-embed-components/src/embed/link-card/types.ts
