+++ packages/zenn-embed-components/src/embed/mermaid/index.tsx
