+++ packages/zenn-embed-components/src/index.ts
@@ -1,5 +1,5 @@
-export * from './components/github';
-export * from './components/gist';
-export * from './components/link-card';
-export * from './components/mermaid';
-export * from './components/tweet';
+export * from './embed/github';
+export * from './embed/gist';
+export * from './embed/link-card';
+export * from './embed/mermaid';
+export * from './embed/tweet';
